fix migrations

+++ db/migrations/20161025142411_cleaningModels.js
@@ -1,24 +1,28 @@
 exports.up = function(knex, Promise) {
     return Promise.all([
-        knex.schema.alterTable('dataset', function(t) {
-            knex.schema.hasColumn('dataset', 'disclaimer').then(exits => {
-                t.dropColumn('disclaimer');
-            })
+        knex.schema.hasColumn('dataset', 'disclaimer').then(exists => {
+            if (exists) {
+                return knex.schema.alterTable('dataset', function(t) {
+                    return t.dropColumn('disclaimer');
+                })
+            }
         }),
-        knex.schema.alterTable('file', function(t) {
-            knex.schema.hasColumn('file', 'collection').then(exits => {
-                t.dropColumn('collection');
-            })
-            knex.schema.hasColumn('file', 'updateDate').then(exits => {
-                t.dropColumn('updateDate');
-            })
+        knex.schema.hasColumn('file', 'collection').then(exists => {
+            if (exists) {
+                return knex.schema.alterTable('file', function(t) {
+                    return t.dropColumn('collection');
+                })
+            }
+        }),
+        knex.schema.hasColumn('file', 'updateDate').then(exists => {
+            if (exists) {
+                return knex.schema.alterTable('file', function(t) {
+                    t.dropColumn('updateDate');
+                })
+            }
         }),
         knex.schema.dropTableIfExists('mimetype')
-
-    ]).catch(function(error) {
-        console.log(error);
-    });
-
+    ])
 };
 
 exports.down = function(knex, Promise) {
